[#47] UnizController 기초 구현 - 오타 수정

+++ src/main/java/com/uniz/service/UnizService.java
@@ -10,7 +10,7 @@ public interface UnizService {
 
 	public boolean registerSelectKey(UnizVO uniz);
 
-	public UnizVO get(int unizSN);
+	public UnizVO get(Long unizSN);
 
 	public UnizVO findByKeywordForUserUnizInsert(String keyword);
